ai-service/requirements.txt
flask
flask-cors
pandas
scikit-learn
requests
python-dotenv
